Fix libdir for distros using e.g. /usr/lib64
I see * Install to $prefix/lib instead of $libdir in NEWS,
but it breaks files layout on some distributions,
e.g. on ROSA _libexecdir = %{_exec_prefix}/%{_lib},
_lib = lib on i586 and _lib = lib64 on x86_64,
so _libexecdir = _libdir.

Arch-dependent binaries must not be located in /usr/lib on x86_64,
because /usr/lib is used for 32 bit binaries on 64 bit systems.